@@ -4,12 +4,9 @@ namespace Neurotic
 {
     public class IPipe
     {
-        protected float val = 0;
-        protected Tuple<float, float> weightBias = new Tuple<float, float>(0,0);
-        public void SetValue(float value) => val = value;
-        public float GetValue() => val;
-        public void SetWeightBias(Tuple<float, float> wb) => weightBias = wb;
-        public Tuple<float, float> GetWeightBais() => weightBias;
+        protected double val = 0;
+        public void SetValue(double value) => val = value;
+        public double GetValue() => val;
 
         public override bool Equals(object obj)
         {
@@ -23,7 +20,7 @@ namespace Neurotic
 
         public override string ToString()
         {
-            return string.Format($"Pipe -- Val:{val} WB:{weightBias.ToString()}");
+            return string.Format($"Pipe -- Val:{val}");
         }
     }
 }
